pyproject.toml
[tool.poetry]
name = "SymbolicAGI"
version = "0.1.0"
description = "Modular, self-improving AGI framework built on symbolic reasoning, cognitive loops, and foot-slave humiliation."
authors = ["Todd <youremail@example.com>"]
packages = [{ include = "symbolic_agi" }]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.30.5"
redis = "^5.0.3"
playwright = "^1.44.0"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pydantic = "^2.7.1"
aiofiles = "^23.2.1"
faiss-cpu = "^1.8.0"
beautifulsoup4 = "^4.13.4"
pyyaml = "^6.0.2"
ddgs = "^9.5.2"
requests = "^2.32.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.4.7"
mypy = "^1.17.1"
pre-commit = "^4.2.0"
pytest-asyncio = "0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
